# backend/docs/swagger.yaml
definitions:
...
  internal_handler.PrepayRequest:
    properties:
      amount:
        description: amount in cents (fen)
        type: integer
      description:
        type: string
      order_no:
        type: string
      payment_type:
        enum:
        - wechat
        - alipay
        type: string
    required:
    - amount
    - description
    - order_no
    - payment_type
    type: object
...
  internal_handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /payments/alipay/callback:
    post:
      description: Handle Alipay payment callback (called by Alipay server)
      produces:
      - text/plain
      responses:
        "200":
          description: success response
          schema:
            type: string
        "400":
          description: error response
          schema:
            type: string
      summary: Alipay callback
      tags:
      - payments
  /payments/alipay/verify:
    get:
      description: Verify Alipay SDK configuration for debugging
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      summary: Verify Alipay SDK
      tags:
      - payments
  /payments/logs:
    get:
      description: Get payment logs for audit
      parameters:
      - description: Filter by order ID
        in: query
        name: order_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.Response'
      summary: Payment logs
      tags:
      - payments
  /payments/prepay:
    post:
      consumes:
      - application/json
      description: Create a unified payment order (WeChat or Alipay) and return client
        parameters
      parameters:
      - description: Payment order info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.PrepayRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.Response'
      security:
      - BearerAuth: []
      summary: Create payment order
      tags:
      - payments
  /payments/wechat/callback:
    post:
      consumes:
      - application/json
      description: Handle WeChat Pay payment callback (called by WeChat server)
      produces:
      - text/plain
      responses:
        "200":
          description: success response
          schema:
            type: string
        "500":
          description: error response
          schema:
            type: string
      summary: WeChat Pay callback
      tags:
      - payments
...
  /upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file (image/document)
      parameters:
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Upload response with URL
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Upload file
      tags:
      - upload
  /upload/{filename}:
    delete:
      description: Delete an uploaded file by filename (requires auth)
      parameters:
      - description: Filename to delete
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Delete response
          schema:
            additionalProperties:
              type: boolean
            type: object
        "400":
          description: Error response
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: File not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete uploaded file
      tags:
      - upload
swagger: "2.0"